refactor: split skill history scan ownership
