+++ packages/gateway-protocol/src/index.ts
@@ -1,4 +1,3 @@
-// Public gateway protocol entrypoint: wire types, schemas, and validators.
 export {
   buildClawHubTrustErrorDetails,
   ClawHubTrustErrorCodes,
@@ -12,6 +11,7 @@ export { formatValidationErrors, type ValidationError } from "./validation-error
 import { lazyCompile } from "./protocol-validator.js";
 export type { ProtocolValidator } from "./protocol-validator.js";
 export * from "./schema/worker-inference.js";
+export * from "./schema/skill-history.js";
 export * from "./migration-api.js";
 export type * from "./public-session-catalog.js";
 import {
@@ -416,9 +416,6 @@ import {
   SkillsProposalActionParamsSchema,
   SkillsProposalApplyResultSchema,
   SkillsProposalCreateParamsSchema,
-  SkillsProposalHistoryScanParamsSchema,
-  SkillsProposalHistoryScanResultSchema,
-  SkillsProposalHistoryStatusParamsSchema,
   SkillsProposalInspectParamsSchema,
   SkillsProposalInspectResultSchema,
   SkillsProposalRecordResultSchema,
@@ -752,12 +749,6 @@ export const validateSkillsDetailParams = lazyCompile(SkillsDetailParamsSchema);
 export const validateSkillsCuratorStatusParams = lazyCompile(SkillsCuratorStatusParamsSchema);
 export const validateSkillsCuratorActionParams = lazyCompile(SkillsCuratorActionParamsSchema);
 export const validateSkillsProposalsListParams = lazyCompile(SkillsProposalsListParamsSchema);
-export const validateSkillsProposalHistoryStatusParams = lazyCompile(
-  SkillsProposalHistoryStatusParamsSchema,
-);
-export const validateSkillsProposalHistoryScanParams = lazyCompile(
-  SkillsProposalHistoryScanParamsSchema,
-);
 export const validateSkillsProposalInspectParams = lazyCompile(SkillsProposalInspectParamsSchema);
 export const validateSkillsProposalCreateParams = lazyCompile(SkillsProposalCreateParamsSchema);
 export const validateSkillsProposalUpdateParams = lazyCompile(SkillsProposalUpdateParamsSchema);
@@ -1182,9 +1173,6 @@ export {
   SkillsDetailResultSchema,
   SkillsProposalsListParamsSchema,
   SkillsProposalsListResultSchema,
-  SkillsProposalHistoryStatusParamsSchema,
-  SkillsProposalHistoryScanParamsSchema,
-  SkillsProposalHistoryScanResultSchema,
   SkillsProposalInspectParamsSchema,
   SkillsProposalInspectResultSchema,
   SkillsProposalCreateParamsSchema,
@@ -1498,9 +1486,6 @@ export type {
   SkillsDetailResult,
   SkillsProposalsListParams,
   SkillsProposalsListResult,
-  SkillsProposalHistoryStatusParams,
-  SkillsProposalHistoryScanParams,
-  SkillsProposalHistoryScanResult,
   SkillsProposalInspectParams,
   SkillsProposalInspectResult,
   SkillsProposalCreateParams,
