+++ packages/gateway-protocol/src/schema/protocol-schemas.ts
@@ -50,9 +50,6 @@ import {
   SkillsProposalActionParamsSchema,
   SkillsProposalApplyResultSchema,
   SkillsProposalCreateParamsSchema,
-  SkillsProposalHistoryScanParamsSchema,
-  SkillsProposalHistoryScanResultSchema,
-  SkillsProposalHistoryStatusParamsSchema,
   SkillsProposalInspectParamsSchema,
   SkillsProposalInspectResultSchema,
   SkillsProposalRecordResultSchema,
@@ -60,8 +57,6 @@ import {
   SkillsProposalRequestRevisionResultSchema,
   SkillsProposalReviseParamsSchema,
   SkillsProposalUpdateParamsSchema,
-  SkillsProposalsListParamsSchema,
-  SkillsProposalsListResultSchema,
   SkillsSearchParamsSchema,
   SkillsSearchResultSchema,
   SkillsSecurityVerdictsParamsSchema,
@@ -441,6 +436,7 @@ import {
   SessionsSendParamsSchema,
   SessionsUsageParamsSchema,
 } from "./sessions.js";
+import { SkillWorkshopProtocolSchemas } from "./skill-protocol-schemas.js";
 import { PresenceEntrySchema, SnapshotSchema, StateVersionSchema } from "./snapshot.js";
 import { SystemInfoParamsSchema, SystemInfoResultSchema } from "./system-info.js";
 import { TalkSessionAcknowledgeMarkParamsSchema } from "./talk-marks.js";
@@ -841,11 +837,7 @@ export const ProtocolSchemas = {
   SkillsCuratorActionResult: SkillsCuratorActionResultSchema,
   SkillsCuratorStatusParams: SkillsCuratorStatusParamsSchema,
   SkillsCuratorStatusResult: SkillsCuratorStatusResultSchema,
-  SkillsProposalsListParams: SkillsProposalsListParamsSchema,
-  SkillsProposalsListResult: SkillsProposalsListResultSchema,
-  SkillsProposalHistoryStatusParams: SkillsProposalHistoryStatusParamsSchema,
-  SkillsProposalHistoryScanParams: SkillsProposalHistoryScanParamsSchema,
-  SkillsProposalHistoryScanResult: SkillsProposalHistoryScanResultSchema,
+  ...SkillWorkshopProtocolSchemas,
   SkillsProposalInspectParams: SkillsProposalInspectParamsSchema,
   SkillsProposalInspectResult: SkillsProposalInspectResultSchema,
   SkillsProposalCreateParams: SkillsProposalCreateParamsSchema,
