+++ packages/gateway-protocol/src/schema/skill-protocol-schemas.ts
@@ -1,0 +1,17 @@
+import {
+  SkillsProposalsListParamsSchema,
+  SkillsProposalsListResultSchema,
+} from "./agents-models-skills.js";
+import {
+  SkillsProposalHistoryScanParamsSchema,
+  SkillsProposalHistoryScanResultSchema,
+  SkillsProposalHistoryStatusParamsSchema,
+} from "./skill-history.js";
+
+export const SkillWorkshopProtocolSchemas = {
+  SkillsProposalsListParams: SkillsProposalsListParamsSchema,
+  SkillsProposalsListResult: SkillsProposalsListResultSchema,
+  SkillsProposalHistoryStatusParams: SkillsProposalHistoryStatusParamsSchema,
+  SkillsProposalHistoryScanParams: SkillsProposalHistoryScanParamsSchema,
+  SkillsProposalHistoryScanResult: SkillsProposalHistoryScanResultSchema,
+} as const;
