+++ src/agents/embedded-agent-runner/run/skill-workshop-attempt-params.ts
@@ -1,0 +1,20 @@
+import type { RunEmbeddedAgentParams } from "./params.js";
+
+export function resolveSkillWorkshopAttemptParams(
+  params: Pick<
+    RunEmbeddedAgentParams,
+    | "skillWorkshopOrigin"
+    | "skillWorkshopProposalEnv"
+    | "skillWorkshopProposalMutationBudget"
+    | "skillWorkshopProposalOnly"
+    | "skillWorkshopProposalReviewCompletion"
+  >,
+) {
+  return {
+    skillWorkshopProposalOnly: params.skillWorkshopProposalOnly,
+    skillWorkshopProposalEnv: params.skillWorkshopProposalEnv,
+    skillWorkshopOrigin: params.skillWorkshopOrigin,
+    skillWorkshopProposalMutationBudget: params.skillWorkshopProposalMutationBudget,
+    skillWorkshopProposalReviewCompletion: params.skillWorkshopProposalReviewCompletion,
+  };
+}
