+++ src/auto-reply/heartbeat-filter.ts
@@ -239,7 +239,7 @@ function collectSuccessfulToolResultCallIds(message: {
   return uniqueStrings(ids);
 }
 
-function isRealNonHeartbeatUserMessage(
+export function isRealNonHeartbeatUserMessage(
   message: { role: string; content?: unknown },
   heartbeatPrompt?: string,
 ): boolean {
@@ -486,37 +486,5 @@ export function filterHeartbeatTranscriptArtifacts<T extends { role: string; con
     i = next;
   }
 
-  return result;
-}
-
-/** Remove complete scheduled heartbeat turns, including visible work, from a shared transcript. */
-export function filterHeartbeatTranscriptTurns<T extends { role: string; content?: unknown }>(
-  messages: readonly T[],
-  heartbeatPrompt?: string,
-): T[] {
-  const result: T[] = [];
-  let index = 0;
-  while (index < messages.length) {
-    const message = expectDefined(messages[index], "messages entry at index");
-    if (!isHeartbeatUserMessage(message, heartbeatPrompt)) {
-      result.push(message);
-      index++;
-      continue;
-    }
-
-    // Heartbeats share the main transcript. Everything through the next real
-    // user turn belongs to the scheduled run, including tool calls and alerts.
-    index++;
-    while (index < messages.length) {
-      const next = expectDefined(messages[index], "messages entry after heartbeat");
-      if (
-        isHeartbeatUserMessage(next, heartbeatPrompt) ||
-        isRealNonHeartbeatUserMessage(next, heartbeatPrompt)
-      ) {
-        break;
-      }
-      index++;
-    }
-  }
   return result;
 }
