+++ src/config/sessions/session-accessor.sqlite.ts
@@ -95,11 +95,23 @@ import type {
   TranscriptMessageAppendResult,
   TranscriptUpdatePayload,
 } from "./session-accessor.sqlite-contract.js";
+import { listSqliteTranscriptInstancesFromDatabase } from "./session-accessor.sqlite-history.js";
+import {
+  createFallbackSessionEntry,
+  normalizeSqliteChatType,
+  normalizeSqliteNumber,
+  normalizeSqliteText,
+} from "./session-accessor.sqlite-normalize.js";
+import {
+  bindSessionEntryProvenance,
+  resolveSessionEntryProvenanceRow,
+} from "./session-accessor.sqlite-provenance.js";
 import {
   normalizeSqliteStatus,
   parseSqliteSessionEntryJson as parseSessionEntryRow,
   readSqliteSessionEntriesByStatus,
 } from "./session-accessor.sqlite-status.js";
+import { preserveSqliteSameKeySessionRolloverLineage } from "./session-entry-lineage.js";
 import { resolveSqliteTargetFromSessionStorePath } from "./session-sqlite-target.js";
 import {
   deleteSessionTranscriptIndexInTransaction,
@@ -443,77 +455,15 @@ export function listSqliteSessionTranscriptInstances(
 ): SessionTranscriptInstance[] {
   const resolved = resolveSqliteScope({ ...scope, sessionKey: "" });
   const database = openOpenClawAgentDatabase(toDatabaseOptions(resolved));
-  const db = getSessionKysely(database.db);
   const currentEntries = new Map(
     listSqliteSessionEntries(scope).map((summary) => [summary.sessionKey, summary.entry]),
   );
-  const rows = executeSqliteQuerySync(
-    database.db,
-    db
-      .selectFrom("sessions")
-      .select([
-        "session_id",
-        "session_key",
-        "transcript_updated_at",
-        "session_entry_provenance",
-        "acp_owned",
-        "plugin_owner_id",
-        "hook_external_content_source",
-        "parent_session_key",
-        "spawned_by",
-        "chat_type",
-      ])
-      .where("transcript_updated_at", "is not", null)
-      .orderBy("transcript_updated_at", "desc")
-      .orderBy("session_id", "asc"),
-  ).rows;
-  const databasePath = resolveOpenClawAgentSqlitePath(toDatabaseOptions(resolved));
-  return rows
-    .map((row): SessionTranscriptInstance | undefined => {
-      if (isInternalSessionEffectsKey(row.session_key)) {
-        return undefined;
-      }
-      if (row.transcript_updated_at === null) {
-        return undefined;
-      }
-      const updatedAtMs = normalizeSqliteNumber(row.transcript_updated_at);
-      const current = currentEntries.get(row.session_key);
-      // A current entry cannot classify transcript content written before v7.
-      // Keep explicit unknown provenance fail-closed even when identities match.
-      const currentIsExact = current?.sessionId === row.session_id;
-      const provenanceKnown = row.session_entry_provenance === 1;
-      const acpOwned = row.acp_owned === 1 || Boolean(currentIsExact && current?.acp);
-      const hookExternalContentSource =
-        row.hook_external_content_source === "gmail" ||
-        row.hook_external_content_source === "webhook"
-          ? row.hook_external_content_source
-          : undefined;
-      const chatType = normalizeSqliteChatType(row.chat_type) ?? undefined;
-      const entry: SessionEntry = {
-        ...(currentIsExact ? cloneSessionEntry(current) : {}),
-        sessionId: row.session_id,
-        sessionFile: formatSqliteSessionFileMarker({
-          agentId: resolved.agentId,
-          sessionId: row.session_id,
-          storePath: databasePath,
-        }),
-        updatedAt: updatedAtMs,
-        ...(row.parent_session_key ? { parentSessionKey: row.parent_session_key } : {}),
-        ...(row.spawned_by ? { spawnedBy: row.spawned_by, spawnDepth: 1 } : {}),
-        ...(chatType ? { chatType } : {}),
-        ...(provenanceKnown && row.plugin_owner_id ? { pluginOwnerId: row.plugin_owner_id } : {}),
-        ...(provenanceKnown && hookExternalContentSource ? { hookExternalContentSource } : {}),
-      };
-      return {
-        acpOwned,
-        entry,
-        provenanceKnown,
-        sessionId: row.session_id,
-        sessionKey: row.session_key,
-        updatedAtMs,
-      };
-    })
-    .filter((entry): entry is SessionTranscriptInstance => entry !== undefined);
+  return listSqliteTranscriptInstancesFromDatabase({
+    agentId: resolved.agentId,
+    currentEntries,
+    database,
+    databasePath: resolveOpenClawAgentSqlitePath(toDatabaseOptions(resolved)),
+  });
 }
 
 /** Reads a session activity timestamp from the additive SQLite session store. */
@@ -2561,15 +2511,6 @@ function normalizeSqliteSessionKey(sessionKey: string): string {
   return normalizeStoreSessionKey(sessionKey);
 }
 
-function createFallbackSessionEntry(patch: Partial<SessionEntry>): SessionEntry {
-  const now = Date.now();
-  return {
-    sessionId: patch.sessionId ?? randomUUID(),
-    updatedAt: patch.updatedAt ?? now,
-    ...patch,
-  };
-}
-
 function cloneSessionEntry(entry: SessionEntry): SessionEntry {
   return structuredClone(entry);
 }
@@ -2732,45 +2673,6 @@ function emitCommittedLifecycleIdentityMutations(params: {
     current.set(upsert.sessionKey, upsert.entry);
   }
   emitCommittedSessionIdentityDiff(previous, current);
-}
-
-function preserveSqliteSameKeySessionRolloverLineage(params: {
-  next: SessionEntry;
-  previous: SessionEntry;
-  sessionKey: string;
-}): SessionEntry {
-  const previousSessionId = params.previous.sessionId.trim();
-  const nextSessionId = params.next.sessionId.trim();
-  if (!previousSessionId || !nextSessionId || previousSessionId === nextSessionId) {
-    return params.next;
-  }
-
-  return {
-    ...params.next,
-    usageFamilyKey:
-      params.next.usageFamilyKey ?? params.previous.usageFamilyKey ?? params.sessionKey,
-    usageFamilySessionIds: uniqueStrings([
-      ...(params.previous.usageFamilySessionIds ?? []),
-      previousSessionId,
-      ...(params.next.usageFamilySessionIds ?? []),
-      nextSessionId,
-    ]),
-  };
-}
-
-function normalizeSqliteText(value: unknown): string | null {
-  return typeof value === "string" && value.trim() ? value.trim() : null;
-}
-
-function normalizeSqliteChatType(value: unknown): "direct" | "group" | "channel" | null {
-  if (value === "direct" || value === "group" || value === "channel") {
-    return value;
-  }
-  return null;
-}
-
-function normalizeSqliteNumber(value: number | bigint): number {
-  return typeof value === "bigint" ? Number(value) : value;
 }
 
 function assertNonMessageTranscriptEvent(event: TranscriptEvent): void {
@@ -4289,32 +4191,6 @@ function writeSessionEntry(
   const normalizedEntry = normalizeSqliteSessionEntryTimestamp(entry);
   const updatedAt = normalizedEntry.updatedAt;
   const previousEntry = readExactSessionEntryRow(database, sessionKey)?.entry;
-  const existingRoot = executeSqliteQueryTakeFirstSync(
-    database.db,
-    db
-      .selectFrom("sessions")
-      .select([
-        "session_entry_provenance",
-        "acp_owned",
-        "plugin_owner_id",
-        "hook_external_content_source",
-      ])
-      .where("session_id", "=", normalizedEntry.sessionId),
-  );
-  const existingTranscriptEvent = executeSqliteQueryTakeFirstSync(
-    database.db,
-    db
-      .selectFrom("transcript_events")
-      .select("seq")
-      .where("session_id", "=", normalizedEntry.sessionId)
-      .limit(1),
-  );
-  // An ordinary update to a migrated current entry still cannot prove the new
-  // hook-provenance field. Only a newly established v7 session starts known.
-  const preserveUnknownProvenance =
-    existingRoot?.session_entry_provenance === 0 &&
-    (previousEntry?.sessionId === normalizedEntry.sessionId ||
-      existingTranscriptEvent !== undefined);
   // Registry writes snapshot the current transcript watermark so recovery can
   // distinguish same-millisecond transcript writes before and after this row.
   const transcriptObservedAt =
@@ -4327,29 +4203,14 @@ function writeSessionEntry(
   });
   const boundSessionRow = {
     ...boundSessionRoot,
-    ...(preserveUnknownProvenance
-      ? {
-          session_entry_provenance: 0,
-          acp_owned: 0,
-          plugin_owner_id: null,
-          hook_external_content_source: null,
-        }
-      : {}),
     transcript_observed_at: transcriptObservedAt,
   };
-  // Exclusion provenance is monotonic for a session. Replacement entries may
-  // omit owner metadata, but that must not make its retained transcript eligible.
-  const sessionRow =
-    existingRoot?.session_entry_provenance === 1
-      ? {
-          ...boundSessionRow,
-          acp_owned: existingRoot.acp_owned === 1 ? 1 : boundSessionRow.acp_owned,
-          plugin_owner_id: boundSessionRow.plugin_owner_id ?? existingRoot.plugin_owner_id,
-          hook_external_content_source:
-            boundSessionRow.hook_external_content_source ??
-            existingRoot.hook_external_content_source,
-        }
-      : boundSessionRow;
+  const sessionRow = resolveSessionEntryProvenanceRow({
+    boundSessionRow,
+    database,
+    entry: normalizedEntry,
+    previousEntry,
+  });
   executeSqliteQuerySync(
     database.db,
     db
@@ -4466,14 +4327,7 @@ function bindSqliteSessionRoot(params: {
     session_scope: resolveSqliteSessionScope(params.entry, params.sessionKey),
     created_at: resolveSqliteSessionCreatedAt(params.entry, updatedAt),
     updated_at: updatedAt,
-    session_entry_provenance: 1,
-    acp_owned: params.entry.acp ? 1 : 0,
-    plugin_owner_id: normalizeSqliteText(params.entry.pluginOwnerId),
-    hook_external_content_source:
-      params.entry.hookExternalContentSource === "gmail" ||
-      params.entry.hookExternalContentSource === "webhook"
-        ? params.entry.hookExternalContentSource
-        : null,
+    ...bindSessionEntryProvenance(params.entry),
     started_at: finiteSqliteNumber(params.entry.startedAt),
     ended_at: finiteSqliteNumber(params.entry.endedAt),
     status: normalizeSqliteStatus(params.entry.status),
