+++ src/config/sessions/session-entry-lineage.ts
@@ -1,0 +1,25 @@
+import { uniqueStrings } from "@openclaw/normalization-core/string-normalization";
+import type { SessionEntry } from "./types.js";
+
+export function preserveSqliteSameKeySessionRolloverLineage(params: {
+  next: SessionEntry;
+  previous: SessionEntry;
+  sessionKey: string;
+}): SessionEntry {
+  const previousSessionId = params.previous.sessionId.trim();
+  const nextSessionId = params.next.sessionId.trim();
+  if (!previousSessionId || !nextSessionId || previousSessionId === nextSessionId) {
+    return params.next;
+  }
+  return {
+    ...params.next,
+    usageFamilyKey:
+      params.next.usageFamilyKey ?? params.previous.usageFamilyKey ?? params.sessionKey,
+    usageFamilySessionIds: uniqueStrings([
+      ...(params.previous.usageFamilySessionIds ?? []),
+      previousSessionId,
+      ...(params.next.usageFamilySessionIds ?? []),
+      nextSessionId,
+    ]),
+  };
+}
