+++ src/config/sessions/session-entry-provenance.ts
@@ -1,0 +1,8 @@
+import type { HookExternalContentSource } from "../../security/external-content.js";
+
+export type SessionEntryProvenance = {
+  /** Plugin id that owns this session through a trusted runtime creation seam. */
+  pluginOwnerId?: string;
+  /** External hook source that has contributed content to this transcript. */
+  hookExternalContentSource?: HookExternalContentSource;
+};
