+++ src/gateway/server-methods/skills-method-names.ts
@@ -1,0 +1,28 @@
+export const SKILLS_GATEWAY_METHOD_NAMES = [
+  "skills.upload.begin",
+  "skills.upload.chunk",
+  "skills.upload.commit",
+  "skills.status",
+  "skills.bins",
+  "skills.search",
+  "skills.detail",
+  "skills.securityVerdicts",
+  "skills.skillCard",
+  "skills.install",
+  "skills.update",
+  "skills.curator.status",
+  "skills.curator.pin",
+  "skills.curator.unpin",
+  "skills.curator.restore",
+  "skills.proposals.list",
+  "skills.proposals.inspect",
+  "skills.proposals.historyStatus",
+  "skills.proposals.historyScan",
+  "skills.proposals.create",
+  "skills.proposals.update",
+  "skills.proposals.revise",
+  "skills.proposals.requestRevision",
+  "skills.proposals.apply",
+  "skills.proposals.reject",
+  "skills.proposals.quarantine",
+] as const;
