+++ src/gateway/server-methods/skills.ts
@@ -11,8 +11,6 @@ import {
   validateSkillsInstallParams,
   validateSkillsProposalActionParams,
   validateSkillsProposalCreateParams,
-  validateSkillsProposalHistoryScanParams,
-  validateSkillsProposalHistoryStatusParams,
   validateSkillsProposalInspectParams,
   validateSkillsProposalRequestRevisionParams,
   validateSkillsProposalReviseParams,
@@ -24,11 +22,6 @@ import {
   validateSkillsStatusParams,
   validateSkillsUpdateParams,
 } from "../../../packages/gateway-protocol/src/index.js";
-import {
-  listAgentIds,
-  resolveAgentWorkspaceDir,
-  resolveDefaultAgentId,
-} from "../../agents/agent-scope.js";
 import { resolveNodeExecEligibility } from "../../agents/exec-defaults.js";
 import { listAgentWorkspaceDirs } from "../../agents/workspace-dirs.js";
 import { redactConfigObject } from "../../config/redact-snapshot.js";
@@ -59,10 +52,6 @@ import {
   unpinCuratedSkill,
 } from "../../skills/workshop/curator.js";
 import {
-  getSkillHistoryScanStatus,
-  runSkillHistoryScan,
-} from "../../skills/workshop/history-scan.js";
-import {
   applySkillProposal,
   inspectSkillProposal,
   listSkillProposals,
@@ -72,14 +61,16 @@ import {
   rejectSkillProposal,
   reviseSkillProposal,
 } from "../../skills/workshop/service.js";
+import { skillProposalHistoryHandlers } from "./skills-proposal-history.js";
 import { skillsUploadHandlers } from "./skills-upload.js";
-import type {
-  GatewayRequestContext,
-  GatewayRequestHandlerOptions,
-  GatewayRequestHandlers,
-  RespondFn,
-} from "./types.js";
-import { assertValidParams, type Validator } from "./validation.js";
+import {
+  resolveSkillsAgentWorkspace,
+  runSkillsProposalWorkspaceHandler,
+  SKILL_PROPOSAL_RESPONSE_HANDLED,
+  type ResolvedSkillsWorkspace,
+} from "./skills-workspace-handler.js";
+import type { GatewayRequestHandlerOptions, GatewayRequestHandlers, RespondFn } from "./types.js";
+import { assertValidParams } from "./validation.js";
 
 type ClawHubInstallResult = Awaited<ReturnType<typeof installSkillFromClawHub>>;
 type ClawHubInstallParams = Parameters<typeof installSkillFromClawHub>[0];
@@ -111,37 +102,6 @@ function installClawHubSkillDeduped(params: ClawHubInstallParams): Promise<ClawH
     .catch(() => undefined);
   return install;
 }
-
-function resolveSkillsAgentWorkspace(params: unknown, context: GatewayRequestContext) {
-  const cfg = context.getRuntimeConfig();
-  const agentIdRaw =
-    params && typeof params === "object" && "agentId" in params
-      ? normalizeOptionalString((params as { agentId?: unknown }).agentId)
-      : undefined;
-  const agentId = agentIdRaw ? normalizeAgentId(agentIdRaw) : resolveDefaultAgentId(cfg);
-  if (agentIdRaw) {
-    // Explicit agent routing must name a configured agent; otherwise a typo
-    // could create or inspect skills under an unintended workspace.
-    const knownAgents = listAgentIds(cfg);
-    if (!knownAgents.includes(agentId)) {
-      return {
-        ok: false as const,
-        error: errorShape(ErrorCodes.INVALID_REQUEST, `unknown agent id "${agentIdRaw}"`),
-      };
-    }
-  }
-  return {
-    ok: true as const,
-    cfg,
-    agentId,
-    workspaceDir: resolveAgentWorkspaceDir(cfg, agentId),
-  };
-}
-
-type ResolvedSkillsWorkspace = Extract<
-  ReturnType<typeof resolveSkillsAgentWorkspace>,
-  { ok: true }
->;
 
 function buildRemoteAwareWorkspaceSkillStatus(resolved: ResolvedSkillsWorkspace) {
   // Remote skill availability depends on the agent's executable-node surface,
@@ -184,37 +144,6 @@ function buildRevisionAgentInstruction(proposal: Awaited<ReturnType<typeof inspe
   ].join("\n");
 }
 
-const SKILL_PROPOSAL_RESPONSE_HANDLED = Symbol("skill proposal response handled");
-
-async function runSkillsProposalWorkspaceHandler<TParams, TResult>(params: {
-  method: string;
-  rawParams: unknown;
-  respond: RespondFn;
-  context: GatewayRequestContext;
-  validate: Validator<TParams>;
-  run: (
-    parsedParams: TParams,
-    resolved: ResolvedSkillsWorkspace,
-  ) => Promise<TResult | typeof SKILL_PROPOSAL_RESPONSE_HANDLED>;
-}): Promise<void> {
-  if (!assertValidParams(params.rawParams, params.validate, params.method, params.respond)) {
-    return;
-  }
-  const resolved = resolveSkillsAgentWorkspace(params.rawParams, params.context);
-  if (!resolved.ok) {
-    params.respond(false, undefined, resolved.error);
-    return;
-  }
-  try {
-    const result = await params.run(params.rawParams, resolved);
-    if (result !== SKILL_PROPOSAL_RESPONSE_HANDLED) {
-      params.respond(true, result, undefined);
-    }
-  } catch (err) {
-    respondSkillWorkshopError(params.respond, err);
-  }
-}
-
 async function forwardSkillWorkshopRevisionToChatSend(
   opts: GatewayRequestHandlerOptions,
   params: {
@@ -252,6 +181,7 @@ async function forwardSkillWorkshopRevisionToChatSend(
 /** Gateway request handlers for skill status, catalogs, installs, updates, and workshop proposals. */
 export const skillsHandlers: GatewayRequestHandlers = {
   ...skillsUploadHandlers,
+  ...skillProposalHistoryHandlers,
   "skills.status": ({ params, respond, context }) => {
     if (!assertValidParams(params, validateSkillsStatusParams, "skills.status", respond)) {
       return;
@@ -440,39 +370,6 @@ export const skillsHandlers: GatewayRequestHandlers = {
       context,
       validate: validateSkillsProposalsListParams,
       run: (_parsedParams, resolved) => listSkillProposals({ workspaceDir: resolved.workspaceDir }),
-    });
-  },
-  "skills.proposals.historyStatus": async ({ params, respond, context }) => {
-    await runSkillsProposalWorkspaceHandler({
-      method: "skills.proposals.historyStatus",
-      rawParams: params,
-      respond,
-      context,
-      validate: validateSkillsProposalHistoryStatusParams,
-      run: (_parsedParams, resolved) =>
-        Promise.resolve(
-          getSkillHistoryScanStatus({
-            agentId: resolved.agentId,
-            config: resolved.cfg,
-            workspaceDir: resolved.workspaceDir,
-          }),
-        ),
-    });
-  },
-  "skills.proposals.historyScan": async ({ params, respond, context }) => {
-    await runSkillsProposalWorkspaceHandler({
-      method: "skills.proposals.historyScan",
-      rawParams: params,
-      respond,
-      context,
-      validate: validateSkillsProposalHistoryScanParams,
-      run: (parsedParams, resolved) =>
-        runSkillHistoryScan({
-          agentId: resolved.agentId,
-          config: resolved.cfg,
-          ...(parsedParams.direction ? { direction: parsedParams.direction } : {}),
-          workspaceDir: resolved.workspaceDir,
-        }),
     });
   },
   "skills.proposals.inspect": async ({ params, respond, context }) => {
