+++ src/skills/workshop/history-scan-review.ts
@@ -1,0 +1,150 @@
+import { randomUUID } from "node:crypto";
+import fs from "node:fs/promises";
+import os from "node:os";
+import path from "node:path";
+import { resolveDefaultModelForAgent } from "../../agents/model-selection-config.js";
+import type { EmbeddedAgentRunResult } from "../../agents/embedded-agent-runner/types.js";
+import type { OpenClawConfig } from "../../config/types.openclaw.js";
+import { CommandLane } from "../../process/lanes.js";
+import { buildSkillHistoryScanPrompt, type SkillHistoryScanPromptSession } from "./history-scan-prompt.js";
+import type { SkillWorkshopProposalReviewProgress } from "./types.js";
+
+export const HISTORY_SCAN_MAX_PROPOSAL_MUTATIONS = 3;
+export const HISTORY_SCAN_SESSION_SEGMENT = "skill-workshop-history-scan";
+const HISTORY_SCAN_TIMEOUT_MS = 10 * 60_000;
+
+export async function runSkillHistoryScanReview(params: {
+  agentId: string;
+  config: OpenClawConfig;
+  env?: NodeJS.ProcessEnv;
+  modelRef?: { model: string; provider: string };
+  onComplete?: (ideasFound: number) => Promise<void>;
+  onProgress?: (progress: SkillWorkshopProposalReviewProgress) => Promise<void>;
+  progress?: SkillWorkshopProposalReviewProgress;
+  runId?: string;
+  sessions: readonly SkillHistoryScanPromptSession[];
+  workspaceDir: string;
+}): Promise<number> {
+  if (params.sessions.length === 0) {
+    return 0;
+  }
+  const modelRef =
+    params.modelRef ?? resolveDefaultModelForAgent({ cfg: params.config, agentId: params.agentId });
+  const proposalMutationBudget = {
+    remaining: params.progress?.remaining ?? HISTORY_SCAN_MAX_PROPOSAL_MUTATIONS,
+    completed: params.progress?.proposalIds.length ?? 0,
+    successfulMutations: params.progress?.successfulMutations ?? 0,
+    failedMutations: 0,
+    mutatedProposalIds: new Set(params.progress?.proposalIds),
+  };
+  const proposalReviewCompletion = params.onComplete
+    ? {
+        completed: false,
+        complete: async () => {
+          const ideasFound = resolveSkillHistoryScanReviewOutcome({
+            ideasFound: proposalMutationBudget.completed,
+            proposalMutationBudgetRemaining: proposalMutationBudget.remaining,
+            successfulMutations: proposalMutationBudget.successfulMutations,
+            failedMutations: proposalMutationBudget.failedMutations,
+          });
+          await params.onComplete?.(ideasFound);
+        },
+        recordProgress: params.onProgress,
+      }
+    : undefined;
+  const tempDir = await fs.mkdtemp(path.join(os.tmpdir(), "openclaw-skill-history-scan-"));
+  const runId = params.runId ?? `${HISTORY_SCAN_SESSION_SEGMENT}:${randomUUID()}`;
+  let runError: unknown;
+  try {
+    const sessionId = randomUUID();
+    const sessionKey = `agent:${params.agentId}:${HISTORY_SCAN_SESSION_SEGMENT}:${sessionId}`;
+    const { runEmbeddedAgent } = await import("../../agents/embedded-agent.js");
+    const result = await runEmbeddedAgent({
+      sessionId,
+      sessionKey,
+      sandboxSessionKey: sessionKey,
+      sessionFile: path.join(tempDir, "session.jsonl"),
+      agentId: params.agentId,
+      trigger: "manual",
+      lane: CommandLane.SkillWorkshopReview,
+      agentHarnessId: "openclaw",
+      agentHarnessRuntimeOverride: "openclaw",
+      workspaceDir: params.workspaceDir,
+      config: params.config,
+      prompt: buildSkillHistoryScanPrompt({
+        sessions: params.sessions,
+        requireCompletion: proposalReviewCompletion !== undefined,
+      }),
+      provider: modelRef.provider,
+      model: modelRef.model,
+      // A smaller configured fallback must not receive a prompt sized for the primary model.
+      modelFallbacksOverride: [],
+      timeoutMs: HISTORY_SCAN_TIMEOUT_MS,
+      runId,
+      toolsAllow: ["skill_workshop"],
+      disableMessageTool: true,
+      disableTrajectory: true,
+      skillWorkshopProposalOnly: true,
+      skillWorkshopProposalEnv: params.env,
+      skillWorkshopProposalMutationBudget: proposalMutationBudget,
+      skillWorkshopProposalReviewCompletion: proposalReviewCompletion,
+      skillWorkshopOrigin: { agentId: params.agentId, runId },
+      cleanupBundleMcpOnRunEnd: true,
+      bootstrapContextMode: "lightweight",
+      skillsSnapshot: { prompt: "", skills: [] },
+      verboseLevel: "off",
+      reasoningLevel: "off",
+      suppressToolErrorWarnings: true,
+    });
+    runError = resolveSkillHistoryScanRunFailure(result);
+  } catch (error) {
+    runError = error;
+  } finally {
+    await fs.rm(tempDir, { recursive: true, force: true });
+  }
+  if (proposalReviewCompletion?.completed) {
+    return proposalMutationBudget.completed;
+  }
+  return resolveSkillHistoryScanReviewOutcome({
+    ideasFound: proposalMutationBudget.completed,
+    proposalMutationBudgetRemaining: proposalMutationBudget.remaining,
+    successfulMutations: proposalMutationBudget.successfulMutations,
+    failedMutations: proposalMutationBudget.failedMutations,
+    ...(runError === undefined ? {} : { runError }),
+  });
+}
+
+export function resolveSkillHistoryScanRunFailure(
+  result: Pick<EmbeddedAgentRunResult, "meta" | "payloads">,
+): Error | undefined {
+  const errorPayload = result.payloads?.find((payload) => payload.isError);
+  const message =
+    result.meta.error?.message.trim() ||
+    result.meta.failureSignal?.message.trim() ||
+    (result.meta.aborted ? "Historical skill scan model run aborted." : undefined) ||
+    errorPayload?.text?.trim();
+  return message || errorPayload
+    ? new Error(message || "Historical skill scan model run failed.")
+    : undefined;
+}
+
+export function resolveSkillHistoryScanReviewOutcome(params: {
+  failedMutations?: number;
+  ideasFound: number;
+  proposalMutationBudgetRemaining: number;
+  successfulMutations: number;
+  runError?: unknown;
+}): number {
+  if (params.runError !== undefined) {
+    throw params.runError;
+  }
+  if ((params.failedMutations ?? 0) > 0) {
+    throw new Error("Historical skill scan has failed proposal mutations to retry.");
+  }
+  const attemptedMutations =
+    HISTORY_SCAN_MAX_PROPOSAL_MUTATIONS - params.proposalMutationBudgetRemaining;
+  if (params.successfulMutations > attemptedMutations) {
+    throw new Error("Historical skill scan proposal accounting is inconsistent.");
+  }
+  return params.ideasFound;
+}
