+++ src/skills/workshop/history-scan.ts
@@ -1,116 +1,72 @@
-import { createHash, randomUUID } from "node:crypto";
-import fs from "node:fs/promises";
-import os from "node:os";
-import path from "node:path";
+import { randomUUID } from "node:crypto";
 import { resolveAgentConfig, resolveAgentDir } from "../../agents/agent-scope.js";
 import { resolveModel } from "../../agents/embedded-agent-runner/model.js";
 import { isEmbeddedAgentRunActive } from "../../agents/embedded-agent-runner/runs.js";
-import type { EmbeddedAgentRunResult } from "../../agents/embedded-agent-runner/types.js";
 import { resolveDefaultModelForAgent } from "../../agents/model-selection-config.js";
-import { filterHeartbeatTranscriptTurns } from "../../auto-reply/heartbeat-filter.js";
 import { resolveHeartbeatPrompt } from "../../auto-reply/heartbeat.js";
 import { resolveStorePath } from "../../config/sessions/paths.js";
 import {
-  listSessionTranscriptInstances,
-  readTranscriptStatsSync,
-  type SessionTranscriptInstance,
-} from "../../config/sessions/session-accessor.js";
-import type { OpenClawConfig } from "../../config/types.openclaw.js";
-import { readSessionMessagesAsync } from "../../gateway/session-transcript-readers.js";
-import { redactSensitiveText } from "../../logging/redact.js";
+  candidateOlderThanCursor,
+  listHistoryScanCandidates,
+  selectSkillHistoryScanCandidates,
+  type SkillHistoryScanCandidate,
+} from "./history-scan-candidates.js";
+import type { SkillHistoryScanPromptSession } from "./history-scan-prompt.js";
 import {
-  createCorePluginStateSyncKeyedStore,
-  MAX_PLUGIN_STATE_ENTRIES_PER_PLUGIN,
-} from "../../plugin-state/plugin-state-store.js";
-import { CommandLane } from "../../process/lanes.js";
+  HISTORY_SCAN_MAX_PROPOSAL_MUTATIONS,
+  HISTORY_SCAN_SESSION_SEGMENT,
+  runSkillHistoryScanReview,
+} from "./history-scan-review.js";
 import {
-  isAcpSessionKey,
-  isCronSessionKey,
-  isSubagentSessionKey,
-} from "../../routing/session-key.js";
-import { formatSkillExperienceReviewTranscript } from "./experience-review-prompt.js";
+  emptyHistoryScanResult,
+  historyScanStateKey,
+  historyScanStore,
+  isStoredHistoryScanState,
+  toPublicHistoryScanResult,
+  withoutPendingHistoryScan,
+  withHistoryScanIdeas,
+  type SkillHistoryScanDirection,
+  type SkillHistoryScanResult,
+  type SkillHistoryScanScope,
+  type StoredSkillHistoryScanSnapshot,
+  type StoredSkillHistoryScanState,
+} from "./history-scan-state.js";
 import {
-  buildSkillHistoryScanPrompt,
-  type SkillHistoryScanPromptSession,
-} from "./history-scan-prompt.js";
+  collectSkillHistoryScanBatch,
+  HISTORY_SCAN_MAX_SESSION_CHARS,
+  HISTORY_SCAN_SESSION_OVERHEAD_CHARS,
+  readHistoryScanSession,
+  resolveSkillHistoryScanTranscriptBudget,
+} from "./history-scan-transcript.js";
 import { getSkillProposalRunProgress } from "./service.js";
 import type { SkillWorkshopProposalReviewProgress } from "./types.js";
 
-const HISTORY_SCAN_SCHEMA = "openclaw.skill-workshop.history-scan.v1";
-const HISTORY_SCAN_MAX_CANDIDATES = 60;
-const HISTORY_SCAN_MAX_SESSIONS = 20;
-const HISTORY_SCAN_MAX_TRANSCRIPT_CHARS = 80_000;
-const HISTORY_SCAN_MAX_SESSION_CHARS = 16_000;
-const HISTORY_SCAN_MAX_RECENT_MESSAGES = 80;
-const HISTORY_SCAN_MAX_LOCAL_TRANSCRIPT_BYTES = 8 * 1024 * 1024;
-const HISTORY_SCAN_MAX_PROPOSAL_MUTATIONS = 3;
-const HISTORY_SCAN_DEFAULT_CONTEXT_TOKENS = 8_192;
-const HISTORY_SCAN_SESSION_OVERHEAD_CHARS = 256;
-const HISTORY_SCAN_MIN_MODEL_ITERATIONS = 6;
-const HISTORY_SCAN_TIMEOUT_MS = 10 * 60_000;
-const HISTORY_SCAN_SESSION_SEGMENT = "skill-workshop-history-scan";
-const HISTORY_SCAN_BLOCKED_SEGMENTS = new Set([
-  "active-memory",
-  "commitments",
-  "heartbeat",
-  "hook",
-  "memory",
-  "skill-workshop-review",
-  HISTORY_SCAN_SESSION_SEGMENT,
-]);
-
-export type SkillHistoryScanDirection = "older" | "newer";
-
-export type SkillHistoryScanResult = {
-  schema: typeof HISTORY_SCAN_SCHEMA;
-  hasScanned: boolean;
-  reviewedSessions: number;
-  ideasFound: number;
-  hasMore: boolean;
-  lastScanReviewed: number;
-  lastScanIdeas: number;
-  lastScanAt?: string;
-  oldestReviewedAt?: string;
-  newestReviewedAt?: string;
-};
-
-export type SkillHistoryScanCursor = {
-  instanceId: string;
-  updatedAtMs: number;
-};
-
-type StoredSkillHistoryScanSnapshot = SkillHistoryScanResult & {
-  oldestCursor?: SkillHistoryScanCursor;
-  newestCursor?: SkillHistoryScanCursor;
-};
-
-type StoredSkillHistoryScanState = StoredSkillHistoryScanSnapshot & {
-  pending?: {
-    direction: SkillHistoryScanDirection;
-    runId: string;
-    next: StoredSkillHistoryScanSnapshot;
-    progress: SkillWorkshopProposalReviewProgress;
-    sessionCursors: SkillHistoryScanCursor[];
-    completed?: {
-      ideasFound: number;
-    };
-  };
-};
-
-export type SkillHistoryScanCandidate = {
-  entry: SessionTranscriptInstance["entry"];
-  instanceId: string;
-  sessionKey: string;
-  updatedAtMs: number;
-};
-
-type SkillHistoryScanRunParams = {
-  agentId: string;
-  config: OpenClawConfig;
-  direction?: SkillHistoryScanDirection;
-  env?: NodeJS.ProcessEnv;
-  workspaceDir: string;
-};
+export {
+  compareSkillHistoryScanCandidates,
+  isSkillHistoryScanSessionEligible,
+  selectSkillHistoryScanCandidates,
+  type SkillHistoryScanCandidate,
+} from "./history-scan-candidates.js";
+export {
+  resolveSkillHistoryScanReviewOutcome,
+  resolveSkillHistoryScanRunFailure,
+  runSkillHistoryScanReview,
+} from "./history-scan-review.js";
+export {
+  getSkillHistoryScanStatus,
+  type SkillHistoryScanCursor,
+  type SkillHistoryScanDirection,
+  type SkillHistoryScanResult,
+} from "./history-scan-state.js";
+export {
+  collectSkillHistoryScanBatch,
+  formatSkillHistoryScanTranscript,
+  hasLegacyHookTranscriptContent,
+  isSkillHistoryScanLocalTranscriptSizeEligible,
+  prepareSkillHistoryScanReviewMessages,
+  resolveSkillHistoryScanTranscriptBudget,
+  selectSkillHistoryScanReviewMessages,
+} from "./history-scan-transcript.js";
 
 type ActiveSkillHistoryScan = {
   direction: SkillHistoryScanDirection;
@@ -119,538 +75,9 @@ type ActiveSkillHistoryScan = {
 
 const historyScansInFlight = new Map<string, ActiveSkillHistoryScan>();
 
-function historyScanStore(env?: NodeJS.ProcessEnv) {
-  return createCorePluginStateSyncKeyedStore<StoredSkillHistoryScanState>({
-    ownerId: "core:skill-workshop",
-    namespace: "history-scan",
-    maxEntries: MAX_PLUGIN_STATE_ENTRIES_PER_PLUGIN,
-    overflowPolicy: "reject-new",
-    ...(env ? { env } : {}),
-  });
-}
-
-function historyScanStateKey(agentId: string, workspaceDir: string, storePath: string): string {
-  const scope = createHash("sha256")
-    .update(`${agentId}\0${path.resolve(workspaceDir)}\0${path.resolve(storePath)}`)
-    .digest("hex");
-  return `${agentId}:${scope}`;
-}
-
-function emptyHistoryScanResult(): SkillHistoryScanResult {
-  return {
-    schema: HISTORY_SCAN_SCHEMA,
-    hasScanned: false,
-    reviewedSessions: 0,
-    ideasFound: 0,
-    hasMore: false,
-    lastScanReviewed: 0,
-    lastScanIdeas: 0,
-  };
-}
-
-function isStoredHistoryScanState(value: unknown): value is StoredSkillHistoryScanState {
-  return Boolean(
-    value &&
-    typeof value === "object" &&
-    !Array.isArray(value) &&
-    (value as { schema?: unknown }).schema === HISTORY_SCAN_SCHEMA,
-  );
-}
-
-function loadHistoryScanState(params: {
-  agentId: string;
-  config: OpenClawConfig;
-  env?: NodeJS.ProcessEnv;
-  workspaceDir: string;
-}): StoredSkillHistoryScanState | undefined {
-  const storePath = resolveStorePath(params.config.session?.store, {
-    agentId: params.agentId,
-    ...(params.env ? { env: params.env } : {}),
-  });
-  const value = historyScanStore(params.env).lookup(
-    historyScanStateKey(params.agentId, params.workspaceDir, storePath),
-  );
-  return isStoredHistoryScanState(value) ? value : undefined;
-}
-
-export function getSkillHistoryScanStatus(params: {
-  agentId: string;
-  config: OpenClawConfig;
-  env?: NodeJS.ProcessEnv;
-  workspaceDir: string;
-}): SkillHistoryScanResult {
-  return toPublicHistoryScanResult(loadHistoryScanState(params) ?? emptyHistoryScanResult());
-}
-
-function toPublicHistoryScanResult(state: StoredSkillHistoryScanState): SkillHistoryScanResult {
-  const {
-    oldestCursor: _oldestCursor,
-    newestCursor: _newestCursor,
-    pending: _pending,
-    ...result
-  } = state;
-  return result;
-}
-
-function withoutPendingHistoryScan(
-  state: StoredSkillHistoryScanState,
-): StoredSkillHistoryScanSnapshot {
-  const { pending: _pending, ...snapshot } = state;
-  return snapshot;
-}
-
-function withHistoryScanIdeas(params: {
-  next: StoredSkillHistoryScanSnapshot;
-  previous: StoredSkillHistoryScanSnapshot;
-  ideasFound: number;
-}): StoredSkillHistoryScanSnapshot {
-  return {
-    ...params.next,
-    ideasFound: params.previous.ideasFound + params.ideasFound,
-    lastScanIdeas: params.ideasFound,
-  };
-}
-
-export function isSkillHistoryScanSessionEligible(
-  summary: Pick<SessionTranscriptInstance, "acpOwned" | "entry" | "provenanceKnown" | "sessionKey">,
-): boolean {
-  const { acpOwned, entry, provenanceKnown, sessionKey } = summary;
-  if (
-    !provenanceKnown ||
-    acpOwned ||
-    !sessionKey.trim() ||
-    !entry.sessionId?.trim() ||
-    entry.spawnedBy ||
-    (entry.spawnDepth ?? 0) > 0 ||
-    entry.pluginOwnerId ||
-    entry.hookExternalContentSource ||
-    isCronSessionKey(sessionKey) ||
-    isSubagentSessionKey(sessionKey) ||
-    isAcpSessionKey(sessionKey)
-  ) {
-    return false;
-  }
-  const segments = sessionKey.toLowerCase().split(":");
-  return !segments.some((segment) => HISTORY_SCAN_BLOCKED_SEGMENTS.has(segment));
-}
-
-export function compareSkillHistoryScanCandidates(
-  left: Pick<SkillHistoryScanCandidate, "instanceId" | "updatedAtMs">,
-  right: Pick<SkillHistoryScanCandidate, "instanceId" | "updatedAtMs">,
-): number {
-  const timestampOrder = right.updatedAtMs - left.updatedAtMs;
-  if (timestampOrder !== 0) {
-    return timestampOrder;
-  }
-  return left.instanceId < right.instanceId ? -1 : left.instanceId > right.instanceId ? 1 : 0;
-}
-
-function candidateOlderThanCursor(
-  candidate: SkillHistoryScanCandidate,
-  cursor: SkillHistoryScanCursor,
-): boolean {
-  return compareSkillHistoryScanCandidates(candidate, cursor) > 0;
-}
-
-function candidateNewerThanCursor(
-  candidate: SkillHistoryScanCandidate,
-  cursor: SkillHistoryScanCursor,
-): boolean {
-  return compareSkillHistoryScanCandidates(candidate, cursor) < 0;
-}
-
-export function selectSkillHistoryScanCandidates(params: {
-  candidates: readonly SkillHistoryScanCandidate[];
-  direction: SkillHistoryScanDirection;
-  oldestCursor?: SkillHistoryScanCursor;
-  newestCursor?: SkillHistoryScanCursor;
-}): SkillHistoryScanCandidate[] {
-  if (params.direction === "newer") {
-    return params.newestCursor
-      ? params.candidates
-          .filter((candidate) => candidateNewerThanCursor(candidate, params.newestCursor!))
-          .toReversed()
-      : [...params.candidates].toReversed();
-  }
-  return params.oldestCursor
-    ? params.candidates.filter((candidate) =>
-        candidateOlderThanCursor(candidate, params.oldestCursor!),
-      )
-    : [...params.candidates];
-}
-
-function listHistoryScanCandidates(params: SkillHistoryScanRunParams): SkillHistoryScanCandidate[] {
-  const storePath = resolveStorePath(params.config.session?.store, {
-    agentId: params.agentId,
-    ...(params.env ? { env: params.env } : {}),
-  });
-  return listSessionTranscriptInstances({
-    agentId: params.agentId,
-    storePath,
-    readConsistency: "latest",
-    hydrateSkillPromptRefs: false,
-    ...(params.env ? { env: params.env } : {}),
-  })
-    .filter(isSkillHistoryScanSessionEligible)
-    .map(({ entry, sessionId, sessionKey, updatedAtMs }) => ({
-      entry,
-      instanceId: sessionId,
-      sessionKey,
-      updatedAtMs,
-    }))
-    .toSorted(compareSkillHistoryScanCandidates);
-}
-
-function countModelIterations(messages: readonly unknown[]): number {
-  return messages.reduce<number>((count, message) => {
-    if (!message || typeof message !== "object" || Array.isArray(message)) {
-      return count;
-    }
-    return count + ((message as { role?: unknown }).role === "assistant" ? 1 : 0);
-  }, 0);
-}
-
-function capSessionTranscript(transcript: string, maxChars: number): string {
-  if (transcript.length <= maxChars) {
-    return transcript;
-  }
-  const omission = "\n\n[older session content omitted]\n\n";
-  if (maxChars <= omission.length) {
-    return transcript.slice(0, maxChars);
-  }
-  const contentBudget = Math.max(0, maxChars - omission.length);
-  const headLength = Math.min(2_000, Math.floor(contentBudget / 2));
-  const head = transcript.slice(0, headLength);
-  const tail = transcript.slice(-(contentBudget - headLength));
-  return `${head}\n\n[older session content omitted]\n\n${tail}`;
-}
-
-export function hasLegacyHookTranscriptContent(messages: readonly unknown[]): boolean {
-  return messages.some((message) => {
-    if (
-      !message ||
-      typeof message !== "object" ||
-      Array.isArray(message) ||
-      (message as { role?: unknown }).role !== "user"
-    ) {
-      return false;
-    }
-    const rendered = formatSkillExperienceReviewTranscript([message]);
-    return (
-      (rendered.includes("<<<EXTERNAL_UNTRUSTED_CONTENT") &&
-        /(?:^|\n)Source: (?:Email|Webhook)(?:\n|$)/.test(rendered)) ||
-      /(?:^|\n)\[cron:[^\]\n]+\](?: |$)/.test(rendered)
-    );
-  });
-}
-
-export function resolveSkillHistoryScanTranscriptBudget(contextTokens?: number): number {
-  const effectiveContextTokens =
-    Number.isFinite(contextTokens) && (contextTokens ?? 0) > 0
-      ? Math.floor(contextTokens as number)
-      : HISTORY_SCAN_DEFAULT_CONTEXT_TOKENS;
-  return Math.min(
-    HISTORY_SCAN_MAX_TRANSCRIPT_CHARS,
-    Math.max(256, Math.floor(effectiveContextTokens * 0.35)),
-  );
-}
-
-export function formatSkillHistoryScanTranscript(
-  messages: readonly unknown[],
-  maxChars: number,
-): string {
-  // Redact the complete structure first. Truncating first can split a PEM or
-  // other multiline secret so the remaining fragment no longer matches.
-  return capSessionTranscript(
-    // Provider-bound history uses mandatory built-in patterns. Operator log
-    // redaction mode and custom pattern replacement cannot weaken this seam.
-    redactSensitiveText(formatSkillExperienceReviewTranscript(messages), { mode: "tools" }),
-    maxChars,
-  );
-}
-
-function filterSkillHistoryScanReviewMessages(
-  messages: readonly unknown[],
-  heartbeatPrompt?: string,
-): readonly unknown[] | undefined {
-  if (hasLegacyHookTranscriptContent(messages)) {
-    return undefined;
-  }
-  const roleMessages = messages.filter((message): message is { role: string; content?: unknown } =>
-    Boolean(
-      message &&
-      typeof message === "object" &&
-      !Array.isArray(message) &&
-      typeof (message as { role?: unknown }).role === "string",
-    ),
-  );
-  return filterHeartbeatTranscriptTurns(roleMessages, heartbeatPrompt);
-}
-
-export function selectSkillHistoryScanReviewMessages(
-  messages: readonly unknown[],
-  heartbeatPrompt?: string,
-): readonly unknown[] | undefined {
-  return prepareSkillHistoryScanReviewMessages(messages, heartbeatPrompt)?.messages;
-}
-
-export function prepareSkillHistoryScanReviewMessages(
-  messages: readonly unknown[],
-  heartbeatPrompt?: string,
-): { messages: readonly unknown[]; modelIterations: number } | undefined {
-  const filtered = filterSkillHistoryScanReviewMessages(messages, heartbeatPrompt);
-  if (!filtered) {
-    return undefined;
-  }
-  return {
-    messages: filtered.slice(-HISTORY_SCAN_MAX_RECENT_MESSAGES),
-    modelIterations: countModelIterations(filtered),
-  };
-}
-
-export function isSkillHistoryScanLocalTranscriptSizeEligible(sizeBytes: number): boolean {
-  return (
-    Number.isFinite(sizeBytes) &&
-    sizeBytes >= 0 &&
-    sizeBytes <= HISTORY_SCAN_MAX_LOCAL_TRANSCRIPT_BYTES
-  );
-}
-
-async function readHistoryScanSession(params: {
-  agentId: string;
-  candidate: SkillHistoryScanCandidate;
-  heartbeatPrompt: string;
-  maxTranscriptChars: number;
-  storePath: string;
-}): Promise<SkillHistoryScanPromptSession | undefined> {
-  const transcriptScope = {
-    agentId: params.agentId,
-    sessionId: params.candidate.entry.sessionId,
-    sessionKey: params.candidate.sessionKey,
-    sessionEntry: params.candidate.entry,
-    storePath: params.storePath,
-  };
-  // Legacy rows may predate explicit hook provenance. Inspect every local turn
-  // before choosing a bounded provider-facing window so old hook payloads can
-  // never age out of the exclusion check.
-  if (
-    !isSkillHistoryScanLocalTranscriptSizeEligible(
-      readTranscriptStatsSync(transcriptScope).sizeBytes,
-    )
-  ) {
-    return undefined;
-  }
-  const allMessages = await readSessionMessagesAsync(transcriptScope, {
-    mode: "full",
-    reason: "Skill Workshop legacy hook provenance check",
-  });
-  const review = prepareSkillHistoryScanReviewMessages(allMessages, params.heartbeatPrompt);
-  if (!review || review.modelIterations < HISTORY_SCAN_MIN_MODEL_ITERATIONS) {
-    return undefined;
-  }
-  const transcript = formatSkillHistoryScanTranscript(review.messages, params.maxTranscriptChars);
-  if (!transcript.trim()) {
-    return undefined;
-  }
-  return {
-    instanceId: params.candidate.instanceId,
-    sessionKey: params.candidate.sessionKey,
-    updatedAt: new Date(params.candidate.updatedAtMs).toISOString(),
-    modelIterations: review.modelIterations,
-    transcript,
-  };
-}
-
-export async function collectSkillHistoryScanBatch(params: {
-  candidates: readonly SkillHistoryScanCandidate[];
-  isSessionActive?: (candidate: SkillHistoryScanCandidate) => boolean;
-  maxTranscriptChars?: number;
-  readSession: (
-    candidate: SkillHistoryScanCandidate,
-  ) => Promise<SkillHistoryScanPromptSession | undefined>;
-}): Promise<{
-  blockedByActive: boolean;
-  considered: SkillHistoryScanCandidate[];
-  sessions: SkillHistoryScanPromptSession[];
-}> {
-  const considered: SkillHistoryScanCandidate[] = [];
-  const sessions: SkillHistoryScanPromptSession[] = [];
-  const maxTranscriptChars = params.maxTranscriptChars ?? HISTORY_SCAN_MAX_TRANSCRIPT_CHARS;
-  let blockedByActive = false;
-  let transcriptChars = 0;
-  for (const candidate of params.candidates.slice(0, HISTORY_SCAN_MAX_CANDIDATES)) {
-    if (params.isSessionActive?.(candidate)) {
-      blockedByActive = true;
-      break;
-    }
-    const session = await params.readSession(candidate);
-    // An active run can claim the session while its transcript is being read.
-    // Stop before advancing the cursor so a later scan sees a stable snapshot.
-    if (params.isSessionActive?.(candidate)) {
-      blockedByActive = true;
-      break;
-    }
-    if (
-      session &&
-      sessions.length > 0 &&
-      transcriptChars + session.transcript.length + HISTORY_SCAN_SESSION_OVERHEAD_CHARS >
-        maxTranscriptChars
-    ) {
-      break;
-    }
-    considered.push(candidate);
-    if (!session) {
-      continue;
-    }
-    sessions.push(session);
-    transcriptChars += session.transcript.length + HISTORY_SCAN_SESSION_OVERHEAD_CHARS;
-    if (sessions.length >= HISTORY_SCAN_MAX_SESSIONS) {
-      break;
-    }
-  }
-  return { blockedByActive, considered, sessions };
-}
-
-export async function runSkillHistoryScanReview(params: {
-  agentId: string;
-  config: OpenClawConfig;
-  env?: NodeJS.ProcessEnv;
-  modelRef?: { model: string; provider: string };
-  onComplete?: (ideasFound: number) => Promise<void>;
-  onProgress?: (progress: SkillWorkshopProposalReviewProgress) => Promise<void>;
-  progress?: SkillWorkshopProposalReviewProgress;
-  runId?: string;
-  sessions: readonly SkillHistoryScanPromptSession[];
-  workspaceDir: string;
-}): Promise<number> {
-  if (params.sessions.length === 0) {
-    return 0;
-  }
-  const modelRef =
-    params.modelRef ?? resolveDefaultModelForAgent({ cfg: params.config, agentId: params.agentId });
-  const proposalMutationBudget = {
-    remaining: params.progress?.remaining ?? HISTORY_SCAN_MAX_PROPOSAL_MUTATIONS,
-    completed: params.progress?.proposalIds.length ?? 0,
-    successfulMutations: params.progress?.successfulMutations ?? 0,
-    failedMutations: 0,
-    mutatedProposalIds: new Set(params.progress?.proposalIds),
-  };
-  const proposalReviewCompletion = params.onComplete
-    ? {
-        completed: false,
-        complete: async () => {
-          const ideasFound = resolveSkillHistoryScanReviewOutcome({
-            ideasFound: proposalMutationBudget.completed,
-            proposalMutationBudgetRemaining: proposalMutationBudget.remaining,
-            successfulMutations: proposalMutationBudget.successfulMutations,
-            failedMutations: proposalMutationBudget.failedMutations,
-          });
-          await params.onComplete?.(ideasFound);
-        },
-        recordProgress: params.onProgress,
-      }
-    : undefined;
-  const tempDir = await fs.mkdtemp(path.join(os.tmpdir(), "openclaw-skill-history-scan-"));
-  const runId = params.runId ?? `${HISTORY_SCAN_SESSION_SEGMENT}:${randomUUID()}`;
-  let runError: unknown;
-  try {
-    const sessionId = randomUUID();
-    const sessionKey = `agent:${params.agentId}:${HISTORY_SCAN_SESSION_SEGMENT}:${sessionId}`;
-    const { runEmbeddedAgent } = await import("../../agents/embedded-agent.js");
-    const result = await runEmbeddedAgent({
-      sessionId,
-      sessionKey,
-      sandboxSessionKey: sessionKey,
-      sessionFile: path.join(tempDir, "session.jsonl"),
-      agentId: params.agentId,
-      trigger: "manual",
-      lane: CommandLane.SkillWorkshopReview,
-      agentHarnessId: "openclaw",
-      agentHarnessRuntimeOverride: "openclaw",
-      workspaceDir: params.workspaceDir,
-      config: params.config,
-      prompt: buildSkillHistoryScanPrompt({
-        sessions: params.sessions,
-        requireCompletion: proposalReviewCompletion !== undefined,
-      }),
-      provider: modelRef.provider,
-      model: modelRef.model,
-      // Keep the prompt budget tied to the selected model. A smaller configured
-      // fallback must not receive a prompt sized for the primary model.
-      modelFallbacksOverride: [],
-      timeoutMs: HISTORY_SCAN_TIMEOUT_MS,
-      runId,
-      toolsAllow: ["skill_workshop"],
-      disableMessageTool: true,
-      disableTrajectory: true,
-      skillWorkshopProposalOnly: true,
-      skillWorkshopProposalEnv: params.env,
-      skillWorkshopProposalMutationBudget: proposalMutationBudget,
-      skillWorkshopProposalReviewCompletion: proposalReviewCompletion,
-      skillWorkshopOrigin: { agentId: params.agentId, runId },
-      cleanupBundleMcpOnRunEnd: true,
-      bootstrapContextMode: "lightweight",
-      skillsSnapshot: { prompt: "", skills: [] },
-      verboseLevel: "off",
-      reasoningLevel: "off",
-      suppressToolErrorWarnings: true,
-    });
-    runError = resolveSkillHistoryScanRunFailure(result);
-  } catch (error) {
-    runError = error;
-  } finally {
-    await fs.rm(tempDir, { recursive: true, force: true });
-  }
-  if (proposalReviewCompletion?.completed) {
-    return proposalMutationBudget.completed;
-  }
-  return resolveSkillHistoryScanReviewOutcome({
-    ideasFound: proposalMutationBudget.completed,
-    proposalMutationBudgetRemaining: proposalMutationBudget.remaining,
-    successfulMutations: proposalMutationBudget.successfulMutations,
-    failedMutations: proposalMutationBudget.failedMutations,
-    ...(runError === undefined ? {} : { runError }),
-  });
-}
-
-export function resolveSkillHistoryScanRunFailure(
-  result: Pick<EmbeddedAgentRunResult, "meta" | "payloads">,
-): Error | undefined {
-  const errorPayload = result.payloads?.find((payload) => payload.isError);
-  const message =
-    result.meta.error?.message.trim() ||
-    result.meta.failureSignal?.message.trim() ||
-    (result.meta.aborted ? "Historical skill scan model run aborted." : undefined) ||
-    errorPayload?.text?.trim();
-  return message || errorPayload
-    ? new Error(message || "Historical skill scan model run failed.")
-    : undefined;
-}
-
-export function resolveSkillHistoryScanReviewOutcome(params: {
-  failedMutations?: number;
-  ideasFound: number;
-  proposalMutationBudgetRemaining: number;
-  successfulMutations: number;
-  runError?: unknown;
-}): number {
-  if (params.runError !== undefined) {
-    throw params.runError;
-  }
-  if ((params.failedMutations ?? 0) > 0) {
-    throw new Error("Historical skill scan has failed proposal mutations to retry.");
-  }
-  const attemptedMutations =
-    HISTORY_SCAN_MAX_PROPOSAL_MUTATIONS - params.proposalMutationBudgetRemaining;
-  if (params.successfulMutations > attemptedMutations) {
-    throw new Error("Historical skill scan proposal accounting is inconsistent.");
-  }
-  return params.ideasFound;
-}
-
 export function resolveSkillHistoryScanHasMore(params: {
   direction: SkillHistoryScanDirection;
-  oldestCursor?: SkillHistoryScanCursor;
+  oldestCursor?: { instanceId: string; updatedAtMs: number };
   candidates: readonly SkillHistoryScanCandidate[];
 }): boolean {
   // A cursorless newer scan follows an empty first scan. Its candidates are
@@ -736,7 +163,7 @@ function toStoredState(params: {
     candidates: params.candidates,
   });
   return {
-    schema: HISTORY_SCAN_SCHEMA,
+    schema: "openclaw.skill-workshop.history-scan.v1",
     hasScanned: true,
     reviewedSessions: (previous?.reviewedSessions ?? 0) + params.sessions.length,
     ideasFound: (previous?.ideasFound ?? 0) + params.ideasFound,
@@ -756,7 +183,7 @@ function toStoredState(params: {
 }
 
 async function runSkillHistoryScanCore(
-  params: SkillHistoryScanRunParams,
+  params: SkillHistoryScanScope,
 ): Promise<SkillHistoryScanResult> {
   const store = historyScanStore(params.env);
   const storePath = resolveStorePath(params.config.session?.store, {
@@ -809,8 +236,8 @@ async function runSkillHistoryScanCore(
   let eligible = selectSkillHistoryScanCandidates({
     candidates,
     direction,
-    ...(previous?.oldestCursor ? { oldestCursor: previous.oldestCursor } : {}),
-    ...(previous?.newestCursor ? { newestCursor: previous.newestCursor } : {}),
+    ...(previous.oldestCursor ? { oldestCursor: previous.oldestCursor } : {}),
+    ...(previous.newestCursor ? { newestCursor: previous.newestCursor } : {}),
   });
   if (resumedPending) {
     const candidatesById = new Map(
@@ -862,8 +289,7 @@ async function runSkillHistoryScanCore(
     HISTORY_SCAN_MAX_SESSION_CHARS,
     Math.max(1, maxTranscriptChars - HISTORY_SCAN_SESSION_OVERHEAD_CHARS),
   );
-  // Heartbeat turns durably use the stable transcript marker. The configured
-  // prompt is only an extra legacy match and may change without hiding old turns.
+  // The configured prompt is only an extra legacy match; the stable marker is authoritative.
   const heartbeatPrompt = resolveHeartbeatPrompt(
     resolveAgentConfig(params.config, params.agentId)?.heartbeat?.prompt ??
       params.config.agents?.defaults?.heartbeat?.prompt,
@@ -917,15 +343,13 @@ async function runSkillHistoryScanCore(
     store.register(stateKey, provisionalNext);
     return provisionalNext;
   }
-
   const runId = resumedPending?.runId ?? `${HISTORY_SCAN_SESSION_SEGMENT}:${randomUUID()}`;
   const progress = resumedPending?.progress ?? {
     proposalIds: [],
     remaining: HISTORY_SCAN_MAX_PROPOSAL_MUTATIONS,
     successfulMutations: 0,
   };
-  // Write the in-progress checkpoint before any proposal can be persisted.
-  // Only the review's explicit final tool call may mark the whole batch complete.
+  // Checkpoint before persistence. Only the explicit final tool call completes the batch.
   store.register(stateKey, {
     ...previous,
     pending: {
@@ -960,10 +384,7 @@ async function runSkillHistoryScanCore(
         }
         store.register(stateKey, {
           ...previous,
-          pending: {
-            ...current.pending,
-            progress: nextProgress,
-          },
+          pending: { ...current.pending, progress: nextProgress },
         });
       },
       onComplete: async (ideasFound) => {
@@ -977,10 +398,7 @@ async function runSkillHistoryScanCore(
         }
         store.register(stateKey, {
           ...previous,
-          pending: {
-            ...current.pending,
-            completed: { ideasFound },
-          },
+          pending: { ...current.pending, completed: { ideasFound } },
         });
       },
       runId,
@@ -1004,13 +422,12 @@ async function runSkillHistoryScanCore(
     store.register(stateKey, next);
     return next;
   }
-  // Leave the in-progress checkpoint intact. A retry reuses its run id,
-  // durable proposal ids, and remaining mutation budget.
+  // Retry reuses its run id, durable proposal ids, and remaining mutation budget.
   throw reviewError ?? new Error("Historical skill scan did not confirm batch completion.");
 }
 
 export function runSkillHistoryScan(
-  params: SkillHistoryScanRunParams,
+  params: SkillHistoryScanScope,
 ): Promise<SkillHistoryScanResult> {
   const storePath = resolveStorePath(params.config.session?.store, {
     agentId: params.agentId,
@@ -1020,12 +437,13 @@ export function runSkillHistoryScan(
   const direction = params.direction ?? "older";
   const active = historyScansInFlight.get(key);
   if (active) {
-    if (active.direction === direction) {
-      return active.run;
-    }
-    return Promise.reject(
-      new Error(`A Skill Workshop history scan in the ${active.direction} direction is running.`),
-    );
+    return active.direction === direction
+      ? active.run
+      : Promise.reject(
+          new Error(
+            `A Skill Workshop history scan in the ${active.direction} direction is running.`,
+          ),
+        );
   }
   const run = runSkillHistoryScanCore({ ...params, direction }).then(toPublicHistoryScanResult);
   const current = { direction, run };
