+++ src/skills/workshop/service.ts
@@ -1,10 +1,8 @@
 import fs from "node:fs/promises";
 import path from "node:path";
-import { expectDefined } from "@openclaw/normalization-core";
 import { normalizeOptionalString } from "@openclaw/normalization-core/string-coerce";
 import type { OpenClawConfig } from "../../config/types.openclaw.js";
 import { readLocalFileSafely, root, walkDirectory } from "../../infra/fs-safe.js";
-import { normalizeSkillIndexName } from "../discovery/skill-index.js";
 import {
   buildWorkspaceSkillStatus,
   resolveSkillStatusEntry,
@@ -29,15 +27,18 @@ import {
 } from "./frontmatter.js";
 import { assertProposalContainsNoLiteralSecrets, scanProposalBundle } from "./proposal-scan.js";
 import {
+  isProposalInWorkspace,
+  listSkillProposals,
+  readRequiredProposal,
+} from "./service-query.js";
+import {
   createSkillProposalId,
   createSkillProposalRollback,
   hashSkillProposalContent,
   MAX_PROPOSAL_SUPPORT_FILES,
   prepareSkillProposalSupportFiles,
   readProposalSupportFiles,
-  readSkillProposal,
   readSkillProposalRecord,
-  readSkillProposalManifest,
   replaceSkillProposalDraft,
   refreshSkillProposalManifest,
   resolveSkillProposalTarget,
@@ -47,6 +48,13 @@ import {
   withSkillProposalTargetLock,
   type PreparedSkillProposalSupportFile,
 } from "./store.js";
+export {
+  getSkillProposalRunProgress,
+  inspectSkillProposal,
+  listSkillProposalIdsForRun,
+  listSkillProposals,
+  resolvePendingSkillProposal,
+} from "./service-query.js";
 import {
   MAX_SKILL_PROPOSAL_ORIGIN_RUN_IDS,
   SKILL_WORKSHOP_SCHEMA,
@@ -69,11 +77,6 @@ type SkillWorkshopWorkspaceOptions = {
   agentId?: string;
 };
 
-type SkillProposalScopeOptions = {
-  env?: NodeJS.ProcessEnv;
-  workspaceDir?: string;
-};
-
 function proposalStoreOptions(env?: NodeJS.ProcessEnv) {
   return env ? { env } : {};
 }
@@ -82,54 +85,6 @@ const WRITABLE_WORKSPACE_SOURCES = new Set(["openclaw-workspace", "agents-skills
 const MAX_PROPOSAL_DRAFT_BYTES = 1024 * 1024;
 const MAX_PROPOSAL_DIRECTORY_ENTRIES = MAX_PROPOSAL_SUPPORT_FILES * 4;
 const MAX_SKILL_PROPOSAL_DESCRIPTION_BYTES = 160;
-
-/** Lists skill workshop proposals, optionally scoped to a workspace. */
-export async function listSkillProposals(
-  options: SkillProposalScopeOptions = {},
-): Promise<SkillProposalManifest> {
-  const store = options.env ? { env: options.env } : {};
-  const manifest = await readSkillProposalManifest(store);
-  if (!options.workspaceDir) {
-    return manifest;
-  }
-  const proposals: SkillProposalManifest["proposals"] = [];
-  for (const proposal of manifest.proposals) {
-    const record = await readSkillProposalRecord(proposal.id, store);
-    if (record && isProposalInWorkspace(record, options.workspaceDir)) {
-      proposals.push(proposal);
-    }
-  }
-  return { ...manifest, proposals };
-}
-
-/** Lists proposals durably attributed to a review run, including later revisions. */
-export async function listSkillProposalIdsForRun(
-  options: SkillProposalScopeOptions & { runId: string },
-): Promise<string[]> {
-  return (await getSkillProposalRunProgress(options)).proposalIds;
-}
-
-export async function getSkillProposalRunProgress(
-  options: SkillProposalScopeOptions & { runId: string },
-): Promise<{ mutationCount: number; proposalIds: string[] }> {
-  const store = options.env ? { env: options.env } : {};
-  // A proposal record lands before its derived manifest entry. Rebuild from
-  // records so crash recovery cannot miss a durable mutation.
-  const manifest = await refreshSkillProposalManifest(store);
-  const ids: string[] = [];
-  let mutationCount = 0;
-  for (const proposal of manifest.proposals) {
-    const record = await readSkillProposalRecord(proposal.id, store);
-    if (!record || (options.workspaceDir && !isProposalInWorkspace(record, options.workspaceDir))) {
-      continue;
-    }
-    if (record.origin?.runId === options.runId || record.originRunIds?.includes(options.runId)) {
-      ids.push(record.id);
-      mutationCount += record.originRunMutationCounts?.[options.runId] ?? 1;
-    }
-  }
-  return { mutationCount, proposalIds: ids };
-}
 
 export async function readSkillProposalDraftFile(filePath: string): Promise<string> {
   const read = await readLocalFileSafely({
@@ -244,68 +199,6 @@ function mergeProposalOriginRunProvenance(
     ...(ids.size > 0 ? { originRunIds: [...ids] } : {}),
     ...(Object.keys(counts).length > 0 ? { originRunMutationCounts: counts } : {}),
   };
-}
-
-export async function inspectSkillProposal(
-  proposalId: string,
-  options: SkillProposalScopeOptions = {},
-): Promise<SkillProposalReadResult | null> {
-  const read = await readSkillProposal(proposalId, proposalStoreOptions(options.env));
-  if (!read) {
-    return null;
-  }
-  if (options.workspaceDir && !isProposalInWorkspace(read.record, options.workspaceDir)) {
-    return null;
-  }
-  return await hydrateProposalSupportFiles(read, options.env);
-}
-
-export async function resolvePendingSkillProposal(input: {
-  env?: NodeJS.ProcessEnv;
-  proposalId?: string;
-  name?: string;
-  workspaceDir?: string;
-}): Promise<SkillProposalReadResult> {
-  const proposalId = normalizeOptionalString(input.proposalId);
-  if (proposalId) {
-    const direct = await readRequiredProposal(proposalId, input.workspaceDir, input.env);
-    if (direct.record.status !== "pending") {
-      throw new Error(
-        `Only pending proposals can be revised. Current status: ${direct.record.status}.`,
-      );
-    }
-    return direct;
-  }
-
-  const name = normalizeOptionalString(input.name);
-  if (!name) {
-    throw new Error("proposal_id or name required.");
-  }
-  const manifest = await listSkillProposals({ workspaceDir: input.workspaceDir, env: input.env });
-  const matches = manifest.proposals.filter(
-    (proposal) => proposal.status === "pending" && proposalMatchesName(proposal, name),
-  );
-  if (matches.length === 0) {
-    throw new Error(`No pending skill proposal matched: ${name}`);
-  }
-  if (matches.length > 1) {
-    const candidates = matches
-      .slice(0, 8)
-      .map((proposal) => `${proposal.id} (${proposal.skillKey})`)
-      .join(", ");
-    throw new Error(`Multiple pending skill proposals matched ${name}: ${candidates}`);
-  }
-  const matched = await readRequiredProposal(
-    expectDefined(matches[0], "matches capture group 0").id,
-    input.workspaceDir,
-    input.env,
-  );
-  if (matched.record.status !== "pending") {
-    throw new Error(
-      `Only pending proposals can be revised. Current status: ${matched.record.status}.`,
-    );
-  }
-  return matched;
 }
 
 export async function proposeCreateSkill(
@@ -1006,45 +899,6 @@ async function assertSupportTargetsUnchanged(
   }
 }
 
-async function readRequiredProposal(
-  proposalId: string,
-  workspaceDir?: string,
-  env?: NodeJS.ProcessEnv,
-): Promise<SkillProposalReadResult> {
-  const read = await readSkillProposal(proposalId, proposalStoreOptions(env));
-  if (!read || (workspaceDir && !isProposalInWorkspace(read.record, workspaceDir))) {
-    throw new Error(`Skill proposal not found: ${proposalId}`);
-  }
-  return read;
-}
-
-async function hydrateProposalSupportFiles(
-  read: SkillProposalReadResult,
-  env?: NodeJS.ProcessEnv,
-): Promise<SkillProposalReadResult> {
-  const supportFiles = await readProposalSupportFiles(read.record, proposalStoreOptions(env));
-  if (supportFiles.length === 0) {
-    return read;
-  }
-  return {
-    ...read,
-    supportFiles: supportFiles.map((file) => ({
-      path: file.path,
-      content: file.content,
-    })),
-  };
-}
-
-function isProposalInWorkspace(record: SkillProposalRecord, workspaceDir: string): boolean {
-  try {
-    assertInsideWorkspace(workspaceDir, record.target.skillFile, "skill file");
-    assertInsideWorkspace(workspaceDir, record.target.skillDir, "skill directory");
-    return true;
-  } catch {
-    return false;
-  }
-}
-
 async function markProposalStale(
   record: SkillProposalRecord,
   reason: string,
@@ -1058,36 +912,6 @@ async function markProposalStale(
     statusReason: reason,
   };
   await updateSkillProposalRecord({ record: stale, store: proposalStoreOptions(env) });
-}
-
-function proposalMatchesName(
-  proposal: SkillProposalManifest["proposals"][number],
-  name: string,
-): boolean {
-  const normalizedName = normalizeSkillIndexName(name);
-  const candidates = [
-    proposal.id,
-    proposal.skillName,
-    proposal.skillKey,
-    proposal.title,
-    proposal.description,
-  ];
-  return candidates.some((candidate) => {
-    if (!candidate) {
-      return false;
-    }
-    if (candidate === name || candidate.toLowerCase() === name.toLowerCase()) {
-      return true;
-    }
-    const normalizedCandidate = normalizeSkillIndexName(candidate);
-    return (
-      Boolean(normalizedName) &&
-      Boolean(normalizedCandidate) &&
-      (normalizedCandidate === normalizedName ||
-        normalizedCandidate.includes(normalizedName) ||
-        normalizedName.includes(normalizedCandidate))
-    );
-  });
 }
 
 function assertWritableSkillTarget(workspaceDir: string, skill: SkillStatusEntry): void {
