+++ src/skills/workshop/store.ts
@@ -15,8 +15,8 @@ import {
   MAX_WORKSPACE_SKILL_SUPPORT_FILE_BYTES,
   normalizeWorkspaceSkillSupportPath,
 } from "../lifecycle/workspace-skill-write.js";
+import { hasValidProposalOriginProvenance } from "./proposal-origin-validation.js";
 import {
-  MAX_SKILL_PROPOSAL_ORIGIN_RUN_IDS,
   SKILL_WORKSHOP_MANIFEST_SCHEMA,
   SKILL_WORKSHOP_ROLLBACK_SCHEMA,
   SKILL_WORKSHOP_SCHEMA,
@@ -458,9 +458,7 @@ function parseSkillProposalRecord(raw: unknown): SkillProposalRecord | null {
     typeof record.updatedAt !== "string" ||
     typeof record.draftHash !== "string" ||
     record.draftFile !== PROPOSAL_DRAFT_FILE ||
-    !isValidProposalOrigin(record.origin) ||
-    !isValidProposalOriginRunIds(record.originRunIds) ||
-    !isValidProposalOriginRunMutationCounts(record.originRunMutationCounts, record.originRunIds) ||
+    !hasValidProposalOriginProvenance(record) ||
     !isValidSupportFileList(record.supportFiles) ||
     !record.target ||
     typeof record.target !== "object" ||
@@ -474,62 +472,6 @@ function parseSkillProposalRecord(raw: unknown): SkillProposalRecord | null {
     return null;
   }
   return record;
-}
-
-function isValidProposalOriginRunMutationCounts(value: unknown, originRunIds: unknown): boolean {
-  if (value === undefined) {
-    return true;
-  }
-  if (!value || typeof value !== "object" || Array.isArray(value)) {
-    return false;
-  }
-  const allowedIds = new Set(Array.isArray(originRunIds) ? originRunIds : []);
-  const entries = Object.entries(value);
-  return (
-    entries.length <= MAX_SKILL_PROPOSAL_ORIGIN_RUN_IDS &&
-    entries.every(
-      ([runId, count]) =>
-        Boolean(runId.trim()) &&
-        allowedIds.has(runId) &&
-        typeof count === "number" &&
-        Number.isSafeInteger(count) &&
-        count > 0,
-    )
-  );
-}
-
-function isValidProposalOriginRunIds(value: unknown): boolean {
-  if (value === undefined) {
-    return true;
-  }
-  if (!Array.isArray(value) || value.length > MAX_SKILL_PROPOSAL_ORIGIN_RUN_IDS) {
-    return false;
-  }
-  const ids = new Set<string>();
-  for (const item of value) {
-    if (typeof item !== "string" || !item.trim() || ids.has(item)) {
-      return false;
-    }
-    ids.add(item);
-  }
-  return true;
-}
-
-function isValidProposalOrigin(value: unknown): boolean {
-  if (value === undefined) {
-    return true;
-  }
-  if (!value || typeof value !== "object" || Array.isArray(value)) {
-    return false;
-  }
-  const origin = value as Record<string, unknown>;
-  for (const key of ["agentId", "sessionKey", "runId", "messageId"]) {
-    const item = origin[key];
-    if (item !== undefined && typeof item !== "string") {
-      return false;
-    }
-  }
-  return true;
 }
 
 function isValidSupportFileList(value: unknown): boolean {
