+++ ui/src/pages/skill-workshop/history-scan.e2e.test.ts
